+++ ataskq/conftest.py
@@ -1,6 +1,10 @@
 from pytest import fixture
 import os
 import pytest
+import copy
+
+from .config import load_config
+from .env import ATASKQ_CONFIG
 
 if os.getenv("_PYTEST_RAISE", "0") != "0":
 
@@ -26,9 +30,9 @@ END $$;
 def drop_pg_tables(conn):
     assert "pg" in conn
     import psycopg2
-    from ataskq.handler.postgresql import from_connection_str
+    from ataskq.handler.postgresql import PostgresqlDBHandler
 
-    connection = from_connection_str(conn)
+    connection = PostgresqlDBHandler.from_connection_str(conn)
     db_conn = psycopg2.connect(
         host=connection.host, database=connection.database, user=connection.user, password=connection.password
     )
@@ -44,18 +48,19 @@ def drop_pg_tables(conn):
 
 
 @fixture
-def conn(tmp_path):
-    conn = os.getenv("ATASKQ_CONNECTION", "sqlite://{tmp_path}/ataskq.db.sqlite3")
+def config(tmp_path):
+    config = load_config(ATASKQ_CONFIG or "test")
+    conn = config["connection"]
     if "sqlite" in conn:
-        conn = conn.format(tmp_path=tmp_path)
+        config["connection"] = conn.format(tmp_path=tmp_path)
     elif "pg" in conn:
         # connect and clear all db tables
         drop_pg_tables(conn)
     elif "http" in conn:
-        server_conn = os.getenv("ATASKQ_SERVER_CONNECTION", "pg://postgres:postgres@localhost/postgres")
-        assert "pg" in server_conn, "rest api test must be with posgres server"
+        server_conn = os.getenv("TEST_SERVER_CONNECTION", "pg://postgres:postgres@localhost/postgres")
+        assert "pg" in server_conn, "rest apigi test must be with posgres server"
         drop_pg_tables(server_conn)
     else:
         raise Exception(f"Unkown connection format '{conn}'")
 
-    return conn
+    return copy.deepcopy(config)