+++ ataskq/handler/db_handler.py
@@ -1,13 +1,10 @@
 from datetime import datetime, timedelta
-from typing import List, Tuple
-from pathlib import Path
-from io import TextIOWrapper
+from typing import List
 from abc import abstractmethod
 from datetime import datetime
 
 from .handler import Handler, get_query_kwargs
 from ..imodel import IModel
-from ..env import ATASKQ_DB_INIT_ON_HANDLER_INIT, ATASKQ_DB_LIMIT_DEFAULT
 from .. import __schema_version__
 
 CONNECTION_LOG = [None]
@@ -92,10 +89,9 @@ def expand_query_str(query_str, _where=None, _order_by=None, _limit=None, _offse
 
 
 class DBHandler(Handler):
-    def __init__(self, init_db=ATASKQ_DB_INIT_ON_HANDLER_INIT, logger=None) -> None:
-        super().__init__(logger)
-
-        if init_db:
+    def __init__(self, **kwargs) -> None:
+        super().__init__(**kwargs)
+        if self.config["handler"]["db_init"]:
             self.init_db()
 
     @property
@@ -306,9 +302,9 @@ class DBHandler(Handler):
         )
 
     @transaction_decorator()
-    def count_query(
-        self, c, model_cls: IModel, _where: str = None, _limit: int = ATASKQ_DB_LIMIT_DEFAULT, _offset: int = 0
-    ):
+    def count_query(self, c, model_cls: IModel, _where: str = None, _limit: int = None, _offset: int = 0):
+        if _limit is None:
+            _limit = self.config["api"]["limit"]
         query_str = f"SELECT COUNT(*) FROM {model_cls.table_key()}"
         query_str = expand_query_str(query_str, _where=_where, _limit=_limit, _offset=_offset)
 
@@ -324,9 +320,11 @@ class DBHandler(Handler):
         model_cls: IModel,
         _where: str = None,
         _order_by=None,
-        _limit: int = ATASKQ_DB_LIMIT_DEFAULT,
+        _limit: int = None,
         _offset: int = 0,
     ):
+        if _limit is None:
+            _limit = self.config["api"]["limit"]
         query_str = f"SELECT * FROM {model_cls.table_key()}"
         if _order_by is None:
             _order_by = f"{model_cls.table_key()}.{model_cls.id_key()} ASC"
@@ -343,10 +341,10 @@ class DBHandler(Handler):
     ##################
 
     @transaction_decorator(exclusive=True)
-    def take_next_task(self, c, job_id, level_start, level_stop):
+    def take_next_task(self, c, job_id: int = None, level_start: int = None, level_stop: int = None):
         # imported here to avoid circular dependency
         from ..models import Task, EStatus
-        from .handler import Handler, EAction
+        from .handler import EAction
 
         # todo: add FOR UPDATE in the queries postgresql
         level_query = ""
@@ -356,10 +354,14 @@ class DBHandler(Handler):
             level_query += f" AND level < {level_stop}"
 
         # get pending task with minimum level
+        job_query = ""
+        if job_id is not None:
+            job_query += f" AND job_id = {job_id}"
+
         query = (
-            f"SELECT * FROM tasks WHERE job_id = {job_id} AND status IN ('{EStatus.PENDING}'){level_query} AND level = "
-            f"(SELECT MIN(level) FROM tasks WHERE job_id = {job_id} AND status IN ('{EStatus.PENDING}'){level_query})"
-            f" {self.for_update}"
+            f"SELECT * FROM tasks WHERE status IN ('{EStatus.PENDING}'){job_query}{level_query} AND level = "
+            f"(SELECT MIN(level) FROM tasks WHERE status IN ('{EStatus.PENDING}'){job_query}{level_query})"
+            f" ORDER BY job_id ASC, task_id ASC {self.for_update}"
         )
         query = query.strip()
 
@@ -373,8 +375,8 @@ class DBHandler(Handler):
 
         # get running task with minimum level
         query = (
-            f"SELECT * FROM tasks WHERE job_id = {job_id} AND status IN ('{EStatus.RUNNING}'){level_query} AND level = "
-            f"(SELECT MIN(level) FROM tasks WHERE job_id = {job_id} AND status IN ('{EStatus.RUNNING}'){level_query})"
+            f"SELECT * FROM tasks WHERE status IN ('{EStatus.RUNNING}'){job_query}{level_query} AND level = "
+            f"(SELECT MIN(level) FROM tasks WHERE status IN ('{EStatus.RUNNING}'){job_query}{level_query})"
             f" {self.for_update}"
         )
         query = query.strip()
@@ -433,12 +435,14 @@ class DBHandler(Handler):
         self,
         c,
         job_id,
-        _where: str = None,
         _order_by: str = None,
-        _limit: int = ATASKQ_DB_LIMIT_DEFAULT,
+        _limit: int = None,
         _offset: int = 0,
     ):
         from ..models import EStatus
+
+        if _limit is None:
+            _limit = self.config["api"]["limit"]
 
         query_str = (
             "SELECT level, name,"
@@ -461,10 +465,11 @@ class DBHandler(Handler):
         return ret
 
     @transaction_decorator()
-    def jobs_status(
-        self, c, _where: str = None, _order_by: str = None, _limit: int = ATASKQ_DB_LIMIT_DEFAULT, _offset: int = 0
-    ):
+    def jobs_status(self, c, _order_by: str = None, _limit: int = None, _offset: int = 0):
         from ..models import EStatus
+
+        if _limit is None:
+            _limit = self.config["api"]["limit"]
 
         query_str = (
             "SELECT jobs.job_id, jobs.name, jobs.description, jobs.priority, "