+++ ataskq/logger.py
@@ -1,9 +1,15 @@
+from typing import Union
 import logging
 
 
 class Logger:
-    def __init__(self, logger: logging.Logger or None):
-        self._logger = logger or logging.getLogger("ataskq")
+    def __init__(self, logger: Union[str, logging.Logger, None]):
+        if logger is None:
+            self._logger = logging.getLogger("ataskq")
+        elif isinstance(logger, str):
+            self._logger = logging.getLogger(logger)
+        else:
+            self._logger = logger
 
     def exception(self, *args, **kwargs):
         self._logger.exception(*args, **kwargs)