+++ ataskq/test_ataskq.py
@@ -4,22 +4,35 @@ from copy import copy
 
 import pytest
 
-from ataskq import TaskQ, StateKWArg, Job, Task, targs, EStatus
-from ataskq.handler import DBHandler
-from ataskq.handler import EAction, from_connection_str
+from . import TaskQ, StateKWArg, Job, Task, targs, EStatus
+from .handler import DBHandler
+from .handler import EAction, from_config
 
 from .tasks_utils import dummy_args_task
 
 
+def non_decreasing(L):
+    return all(x <= y for x, y in zip(L, L[1:]))
+
+
+def non_increasing(L):
+    return all(x >= y for x, y in zip(L, L[1:]))
+
+
+def monotonic(L):
+    return non_decreasing(L) or non_increasing(L)
+
+
 @pytest.fixture
-def jtaskq(conn) -> TaskQ:
-    return TaskQ(conn=conn).create_job()
+def jtaskq(config) -> TaskQ:
+    return TaskQ(config=config).create_job()
 
 
-def test_create_job(conn):
-    taskq = TaskQ(conn=conn).create_job()
+def test_create_job(config):
+    taskq = TaskQ(config=config).create_job()
     assert isinstance(taskq, TaskQ)
 
+    conn = config["connection"]
     if "sqlite" in conn:
         assert Path(taskq.handler.db_path).exists()
         assert Path(taskq.handler.db_path).is_file()
@@ -31,21 +44,21 @@ def test_create_job(conn):
         raise Exception(f"unknown db type in connection string '{conn}'")
 
 
-def test_job_default_name(conn):
-    job = TaskQ(conn=conn).create_job().job
+def test_job_default_name(config):
+    job = TaskQ(config=config).create_job().job
     assert job.name is None
 
 
-def test_job_custom_name(conn):
-    job = TaskQ(conn=conn).create_job(name="my_job").job
+def test_job_custom_name(config):
+    job = TaskQ(config=config).create_job(name="my_job").job
     assert job.name == "my_job"
 
 
-def test_task_job_delete_cascade(conn):
+def test_task_job_delete_cascade(config):
     # test that deleting a job deletes all its tasks
-    handler = from_connection_str(conn=conn)
-    taskq1: TaskQ = TaskQ(conn=handler).create_job(name="job1")
-    taskq2: TaskQ = TaskQ(conn=handler).create_job(name="job2")
+    handler = from_config(config)
+    taskq1: TaskQ = TaskQ(handler=handler).create_job(name="job1")
+    taskq2: TaskQ = TaskQ(handler=handler).create_job(name="job2")
     assert Job.count_all(_handler=handler) == 2
 
     taskq1.add_tasks(
@@ -78,11 +91,11 @@ def test_task_job_delete_cascade(conn):
     assert Task.count_all(_handler=handler) == 0
 
 
-def test_state_kwargs_job_delete_cascade(conn):
+def test_state_kwargs_job_delete_cascade(config):
     # todo: test should ne under ataskq
-    handler = from_connection_str(conn=conn)
-    taskq1: TaskQ = TaskQ(conn=handler).create_job(name="job1")
-    taskq2: TaskQ = TaskQ(conn=handler).create_job(name="job2")
+    handler = from_config(config)
+    taskq1: TaskQ = TaskQ(handler=handler).create_job(name="job1")
+    taskq2: TaskQ = TaskQ(handler=handler).create_job(name="job2")
     assert len(Job.get_all(_handler=handler)) == 2
 
     taskq1.add_state_kwargs(
@@ -116,7 +129,7 @@ def test_state_kwargs_job_delete_cascade(conn):
 
 
 def test_update_task_start_time(jtaskq):
-    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
 
     jtaskq.add_tasks(
         [
@@ -142,7 +155,7 @@ def test_update_task_start_time(jtaskq):
 
 
 def test_update_task_status(jtaskq):
-    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
 
     jtaskq.add_tasks(
         [
@@ -183,7 +196,7 @@ def test_update_task_status(jtaskq):
     assert task.done_time == now
 
 
-def _compare_task_taken(task1: Task, task2: Task, job_id=None):
+def _compare_tasks(task1: Task, task2: Task, job_id=None):
     if job_id:
         assert task1.job_id == job_id
     assert task1.job_id == task2.job_id
@@ -194,8 +207,32 @@ def _compare_task_taken(task1: Task, task2: Task, job_id=None):
     assert task1.targs == task2.targs
 
 
+def test_get_tasks(jtaskq):
+    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
+    in_task2 = Task(entrypoint=dummy_args_task, level=2, name="task2")
+    in_task3 = Task(entrypoint=dummy_args_task, level=3, name="task3")
+
+    jtaskq.add_tasks(
+        [
+            in_task3,
+            in_task2,
+            in_task1,
+        ]
+    )
+
+    tasks = jtaskq.get_tasks()
+    assert len(tasks) == 3
+    for t in tasks:
+        if t.level == 1:
+            _compare_tasks(in_task1, t)
+        elif t.level == 2:
+            _compare_tasks(in_task2, t)
+        elif t.level == 3:
+            _compare_tasks(in_task3, t)
+
+
 def test_take_next_task_sanity(jtaskq):
-    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
+    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
 
     jtaskq.add_tasks(
         [
@@ -205,12 +242,12 @@ def test_take_next_task_sanity(jtaskq):
 
     action, task = jtaskq._take_next_task(level=None)
     assert action == EAction.RUN_TASK
-    _compare_task_taken(in_task1, task)
+    _compare_tasks(in_task1, task)
 
 
 def test_take_next_task(jtaskq):
-    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="task1")
-    in_task2 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, name="task2")
+    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
+    in_task2 = Task(entrypoint=dummy_args_task, level=2, name="task2")
     in_task3 = copy(in_task1)
 
     jtaskq.add_tasks(
@@ -227,30 +264,30 @@ def test_take_next_task(jtaskq):
     assert action == EAction.RUN_TASK
     assert task.task_id not in tids
     tids.append(task.task_id)
-    _compare_task_taken(in_task1, task)
+    _compare_tasks(in_task1, task)
 
     action, task = jtaskq._take_next_task(level=None)
     assert action == EAction.RUN_TASK
     assert task.task_id not in tids
     tids.append(task.task_id)
-    _compare_task_taken(in_task3, task)  # note in_task3 is copy of in_task1
+    _compare_tasks(in_task3, task)  # note in_task3 is copy of in_task1
 
     action, task = jtaskq._take_next_task(level=None)
     assert action == EAction.WAIT
     assert task is None
 
 
-def test_take_next_task_2_jobs(conn):
+def test_take_next_task_2_jobs(config):
     # todo: test should ne under ataskq
-    handler = from_connection_str(conn=conn)
-    taskq1: TaskQ = TaskQ(conn=handler).create_job(name="job1")
-    taskq2: TaskQ = TaskQ(conn=handler).create_job(name="job2")
+    handler = from_config(config)
+    taskq1: TaskQ = TaskQ(handler=handler).create_job(name="job1")
+    taskq2: TaskQ = TaskQ(handler=handler).create_job(name="job2")
 
-    in_task1 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, name="taska")
-    in_task2 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="taskb")
-    in_task3 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="taskb")
-    in_task4 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, name="taskd")
-    in_task5 = Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, name="taske")
+    in_task1 = Task(entrypoint=dummy_args_task, level=2, name="taska")
+    in_task2 = Task(entrypoint=dummy_args_task, level=1, name="taskb")
+    in_task3 = Task(entrypoint=dummy_args_task, level=1, name="taskb")
+    in_task4 = Task(entrypoint=dummy_args_task, level=1, name="taskd")
+    in_task5 = Task(entrypoint=dummy_args_task, level=2, name="taske")
 
     taskq1.add_tasks(
         [
@@ -277,30 +314,30 @@ def test_take_next_task_2_jobs(conn):
     assert len(Job.get_all(_handler=handler)) == 2
     assert len(taskq2.get_tasks()) == 2
 
-    # db handler 1
+    # taskq 1
     action, task = taskq1._take_next_task(level=None)
     assert action == EAction.RUN_TASK
     assert task.task_id not in tids
     tids.append(task.task_id)
-    _compare_task_taken(in_task2, task, job_id=jid1)
+    _compare_tasks(in_task2, task, job_id=jid1)
     tids.append(task.task_id)
 
     action, task = taskq1._take_next_task(level=None)
     assert action == EAction.RUN_TASK
     assert task.task_id not in tids
     tids.append(task.task_id)
-    _compare_task_taken(in_task3, task, job_id=jid1)  # note in_task3 is copy of in_task1
+    _compare_tasks(in_task3, task, job_id=jid1)  # note in_task3 is copy of in_task1
 
     action, task = taskq1._take_next_task(level=None)
     assert action == EAction.WAIT
     assert task is None
 
-    # db handler 2
+    # taskq 2
     action, task = taskq2._take_next_task(level=None)
     assert action == EAction.RUN_TASK
     assert task.task_id not in tids
     tids.append(task.task_id)
-    _compare_task_taken(in_task4, task, job_id=jid2)
+    _compare_tasks(in_task4, task, job_id=jid2)
     tids.append(task.task_id)
 
     action, task = taskq2._take_next_task(level=None)
@@ -308,34 +345,64 @@ def test_take_next_task_2_jobs(conn):
     assert task is None
 
 
-def test_get_tasks(jtaskq):
-    in_task1 = Task(entrypoint=dummy_args_task, level=1, name="task1")
-    in_task2 = Task(entrypoint=dummy_args_task, level=2, name="task2")
-    in_task3 = Task(entrypoint=dummy_args_task, level=3, name="task3")
+def test_take_next_all_jobs(tmp_path, config):
+    handler = from_config(config)
+    taskq1: TaskQ = TaskQ(handler=handler).create_job(name="job1")
+    taskq2: TaskQ = TaskQ(handler=handler).create_job(name="job2")
 
-    jtaskq.add_tasks(
+    def init_task(name):
+        return Task(entrypoint=dummy_args_task, name=name)
+
+    in_task1 = init_task("job1 - task1")
+    in_task2 = init_task("job1 - task2")
+    taskq1.add_tasks(
         [
-            in_task3,
-            in_task2,
             in_task1,
+            in_task2,
         ]
     )
 
-    tasks = jtaskq.get_tasks()
-    assert len(tasks) == 3
-    for t in tasks:
-        if t.level == 1:
-            _compare_task_taken(in_task1, t)
-        elif t.level == 2:
-            _compare_task_taken(in_task2, t)
-        elif t.level == 3:
-            _compare_task_taken(in_task3, t)
+    in_task4 = init_task("job2 - task4")
+    in_task5 = init_task("job2 - task5")
+    taskq2.add_tasks(
+        [
+            in_task4,
+            in_task5,
+        ]
+    )
+
+    in_task3 = init_task("job1 - task3")
+    taskq1.add_tasks(
+        [
+            in_task3,
+        ]
+    )
+
+    jid1 = taskq1.job_id
+    jid2 = taskq2.job_id
+
+    # sanity check
+    assert len(Job.get_all(_handler=handler)) == 2
+    assert len(taskq1.get_tasks()) == 3
+    assert len(Job.get_all(_handler=handler)) == 2
+    assert len(taskq2.get_tasks()) == 2
+
+    taskq = TaskQ(config=config)
+    in_tasks = [in_task1, in_task2, in_task3, in_task4, in_task5]
+    in_jids = [jid1, jid1, jid1, jid2, jid2]
+    for t, jid in zip(in_tasks, in_jids):
+        action, task = taskq._take_next_task()
+        assert action == EAction.RUN_TASK
+        assert task.task_id == t.task_id
+        _compare_tasks(t, task, job_id=jid)
+
+    assert not monotonic([t.task_id for t in in_tasks])
 
 
-def test_run_default(conn, tmp_path: Path):
+def test_run_default(config, tmp_path: Path):
     filepath = tmp_path / "file.txt"
 
-    taskq = TaskQ(conn=conn).create_job()
+    taskq = TaskQ(config=config).create_job()
 
     taskq.add_tasks(
         [
@@ -351,10 +418,11 @@ def test_run_default(conn, tmp_path: Path):
     assert filepath.read_text() == "task 0\n" "task 1\n" "task 2\n"
 
 
-def test_run_task_raise_exception(conn):
+def test_run_task_raise_exception(config):
     # no exception raised
     try:
-        taskq: TaskQ = TaskQ(conn=conn, run_task_raise_exception=False).create_job()
+        config["run"]["raise_exception"] = False
+        taskq: TaskQ = TaskQ(config=config).create_job()
         taskq.add_tasks(
             [
                 Task(entrypoint="ataskq.tasks_utils.exception_task", targs=targs(message="task failed")),
@@ -365,7 +433,8 @@ def test_run_task_raise_exception(conn):
         assert False, "exception_task raises exception with run_task_raise_exception=False"
 
     # exception raised
-    taskq: TaskQ = TaskQ(conn=conn, run_task_raise_exception=True).create_job()
+    config["run"]["raise_exception"] = True
+    taskq: TaskQ = TaskQ(config=config).create_job()
     taskq.add_tasks(
         [
             Task(entrypoint="ataskq.tasks_utils.exception_task", targs=targs(message="task failed")),
@@ -376,10 +445,10 @@ def test_run_task_raise_exception(conn):
     assert excinfo.value.args[0] == "task failed"
 
 
-def test_run_2_processes(conn, tmp_path: Path):
+def test_run_2_processes(config, tmp_path: Path):
     filepath = tmp_path / "file.txt"
 
-    taskq = TaskQ(conn=conn).create_job()
+    taskq = TaskQ(config=config).create_job()
 
     taskq.add_tasks(
         [
@@ -398,7 +467,7 @@ def test_run_2_processes(conn, tmp_path: Path):
         ]
     )
 
-    taskq.run(num_processes=2)
+    taskq.run(concurrency=2)
 
     assert filepath.exists()
     text = filepath.read_text()
@@ -408,10 +477,10 @@ def test_run_2_processes(conn, tmp_path: Path):
 
 
 @pytest.mark.parametrize("num_processes", [None, 2])
-def test_run_by_level(conn, tmp_path: Path, num_processes: int):
+def test_run_by_level(config, tmp_path: Path, num_processes: int):
     filepath = tmp_path / "file.txt"
 
-    taskq = TaskQ(conn=conn).create_job()
+    taskq = TaskQ(config=config).create_job()
 
     taskq.add_tasks(
         [
@@ -441,14 +510,14 @@ def test_run_by_level(conn, tmp_path: Path, num_processes: int):
     assert taskq.count_pending_tasks_below_level(3) == 4
 
     assert taskq.count_pending_tasks_below_level(1) == 1
-    taskq.run(level=0, num_processes=num_processes)
+    taskq.run(level=0, concurrency=num_processes)
     taskq.count_pending_tasks_below_level(1) == 0
     assert filepath.exists()
     text = filepath.read_text()
     assert "task 0\n" in text
 
     assert taskq.count_pending_tasks_below_level(2) == 2
-    taskq.run(level=1, num_processes=num_processes)
+    taskq.run(level=1, concurrency=num_processes)
     taskq.count_pending_tasks_below_level(2) == 0
     text = filepath.read_text()
     assert "task 0\n" in text
@@ -456,7 +525,7 @@ def test_run_by_level(conn, tmp_path: Path, num_processes: int):
     assert "task 2\n" in text
 
     assert taskq.count_pending_tasks_below_level(3) == 1
-    taskq.run(level=2, num_processes=num_processes)
+    taskq.run(level=2, concurrency=num_processes)
     taskq.count_pending_tasks_below_level(3) == 0
     text = filepath.read_text()
     assert "task 0\n" in text
@@ -465,14 +534,16 @@ def test_run_by_level(conn, tmp_path: Path, num_processes: int):
     assert "task 3\n" in text
 
 
-def test_monitor_pulse_failure(conn):
+def test_monitor_pulse_failure(config):
     # set monitor pulse longer than timeout
-    taskq = TaskQ(conn=conn, monitor_pulse_interval=10, task_pulse_timeout=1.5).create_job()
+    config["monitor"]["pulse_interval"] = 2
+    config["monitor"]["pulse_timeout"] = 1.5
+    taskq = TaskQ(config=config).create_job()
     taskq.add_tasks(
         [
             # reserved keyward for ignored task for testing
             Task(entrypoint="ataskq.skip_run_task", targs=targs("task will fail")),
-            Task(entrypoint="ataskq.tasks_utils.dummy_args_task", targs=targs("task will success")),
+            Task(entrypoint=dummy_args_task, targs=targs("task will success")),
         ]
     )
     start = datetime.now()
@@ -486,25 +557,27 @@ def test_monitor_pulse_failure(conn):
     assert stop - start > timedelta(seconds=1.5)
 
 
-def test_task_wait_timeout(conn):
+def test_task_wait_timeout(config):
     # set monitor pulse longer than timeout
-    taskq = TaskQ(conn=conn, run_task_raise_exception=True, task_wait_timeout=0).create_job()
+    config["run"]["raise_exception"] = True
+    config["run"]["wait_timeout"] = 0
+    taskq = TaskQ(config=config).create_job()
     taskq.add_tasks(
         [
-            Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=1, targs=targs("task will success", sleep=0.2)),
-            Task(entrypoint="ataskq.tasks_utils.dummy_args_task", level=2, targs=targs("task will success")),
+            Task(entrypoint=dummy_args_task, level=1, targs=targs("task will success", sleep=0.2)),
+            Task(entrypoint=dummy_args_task, level=2, targs=targs("task will success")),
         ]
     )
 
     with pytest.raises(Exception) as excinfo:
-        taskq.run(num_processes=2)
+        taskq.run(concurrency=2)
     assert excinfo.value.args[0] == "Some processes failed, see logs for details"
 
 
-def test_run_with_state_kwargs(conn):
+def test_run_with_state_kwargs(config):
     from ataskq.tasks_utils.counter_task import counter_kwarg, counter_task
 
-    taskq = TaskQ(conn=conn, run_task_raise_exception=True).create_job()
+    taskq = TaskQ(config=config).create_job()
 
     taskq.add_state_kwargs(
         [
@@ -521,9 +594,10 @@ def test_run_with_state_kwargs(conn):
     taskq.run()
 
 
-def test_max_jobs(conn):
+def test_max_jobs(config):
     max_jobs = 10
-    taskq = TaskQ(conn=conn, max_jobs=max_jobs)
+    config["db"]["max_jobs"] = max_jobs
+    taskq = TaskQ(config=config)
     if not isinstance(taskq.handler, DBHandler):
         pytest.skip()