+++ ataskq/test_handler.py
@@ -1,19 +1,16 @@
 # todo: add queries order_by tests
 
-from pathlib import Path
-from copy import copy
-from datetime import datetime
-
 import pytest
 
-from .handler import Handler, from_connection_str
+from .config import load_config
+from .handler import Handler, from_config
 from .handler.db_handler import DBHandler, transaction_decorator
 from .handler import register_handler
 
 
 @pytest.fixture
-def handler(conn) -> Handler:
-    handler = from_connection_str(conn)
+def handler(config) -> Handler:
+    handler = from_config(config)
     register_handler("test_handler", handler)
     return handler
 
@@ -44,8 +41,9 @@ def test_invalid_transaction(handler):
     assert "syntax error" in str(excinfo.value)
 
 
-def test_db_format(conn, handler):
+def test_db_format(config, handler):
     assert isinstance(handler, Handler)
+    conn = config["connection"]
 
     if "sqlite" in conn:
         from .handler.sqlite3 import SQLite3DBHandler
@@ -66,17 +64,17 @@ def test_db_format(conn, handler):
 
 def test_db_invalid_format_no_sep():
     with pytest.raises(RuntimeError) as excinfo:
-        from_connection_str(conn=f"sqlite")
+        from_config(config=load_config({"connection": "sqlite"}, environ=False))
     assert "connection must be of format <type>://<connection string>" == str(excinfo.value)
 
 
 def test_db_invalid_format_no_type():
     with pytest.raises(RuntimeError) as excinfo:
-        from_connection_str(conn=f"://ataskq.db")
+        from_config(load_config({"connection": f"://ataskq.db"}, environ=False))
     assert "missing handler type, connection must be of format <type>://<connection string>" == str(excinfo.value)
 
 
 def test_db_invalid_format_no_connectino():
     with pytest.raises(RuntimeError) as excinfo:
-        from_connection_str(conn=f"sqlite://")
+        from_config(load_config({"connection": "sqlite://"}, environ=False))
     assert "missing connection string, connection must be of format <type>://<connection string>" == str(excinfo.value)