+++ contrib/setup.sh
@@ -1,8 +1,12 @@
 #! /bin/bash
 set -e
 SCRIPT_DIR=$( dirname $(readlink -f $0) )
-cd $SCRIPT_DIR/..
 
+cd $SCRIPT_DIR/../client
+npm install
+npm run build
+
+cd $SCRIPT_DIR/..
 python -m pip install --upgrade pip
 pip install -r requirements.txt
 pre-commit install