+++ examples/run_with_state_kwargs.py
@@ -1,8 +1,10 @@
+from pathlib import Path
+
 import context
 from ataskq.tasks_utils.counter_task import counter_kwarg, counter_task
 from ataskq import TaskQ, Task, targs, StateKWArg
 
-taskq = TaskQ(run_task_raise_exception=True).create_job(name="task with state kwargs")
+taskq = TaskQ(run_task_raise_exception=True).create_job(name=Path(__file__).stem)
 
 taskq.add_state_kwargs(
     [